Multi-Screen Planning (only horizontal)
Adding a control to stop following the mouse and return to the initial state smoothly. Note: Depending on the screen, it may be necessary to add a larger limit, as the position update is being done according to the mouse movement , for this reason, controls for these limits have been added for Lively users.rs.