Additional  Blend Filter
Who would have thought that adding a simple filter would be so challenging?
A blend filter has been added to enhance the ambiance, creating a blending effect with the background. Additionally, users of Lyvely now have immense freedom to customize the filter, allowing it to better match any chosen background.